% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_gtfs.R
\name{append_gtfs}
\alias{append_gtfs}
\title{Append GTFS Files From NetPlan}
\usage{
append_gtfs(
  baseline_filepath,
  proposed_filepath,
  deleted_routes,
  save_csv = FALSE,
  save_RDS = FALSE,
  output_folder = NULL
)
}
\arguments{
\item{baseline_filepath}{Character string in quotes. The location of the folder containing the GTFS files of the baseline network.}

\item{proposed_filepath}{Character string in quotes. The location of the folder containing the GTFS files of the proposed network.}

\item{deleted_routes}{Character. List routes to exclude from the final output.}

\item{save_csv}{T/F Do you want to save csv files of the appended GTFS tables?}

\item{save_RDS}{T/F Do you want to save an RDS object of the appended GTFS list object?}

\item{output_folder}{File path of the place to export file exports}
}
\value{
A list object with appended GTFS information. Optionally, exports to csv and RDS.
}
\description{
Append GTFS Files From NetPlan
}
