% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_service_rte_name.R
\name{clean_service_rte_name}
\alias{clean_service_rte_name}
\title{Clean Service Route Names}
\usage{
clean_service_rte_name(df, route_col)
}
\arguments{
\item{df}{The dataframe that contains Service Route Numbers to be converted to Service Route Names}

\item{route_col}{The column with the route number}
}
\value{
column named 'clean_route' with route name
}
\description{
Clean Service Route Names
}
