% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_service_rte_num.R
\name{clean_service_rte_num}
\alias{clean_service_rte_num}
\title{Clean Service Route Numbers}
\usage{
clean_service_rte_num(route_table, netplan_gtfs = FALSE)
}
\arguments{
\item{route_table}{The dataframe representation of routes.txt from your GTFS file}

\item{netplan_gtfs}{T/F Is this GTFS from NetPlan? This impacts how route names are stored and generated}
}
\value{
A route dataframe with a field service_rte_num that representing the route number of Metro routes
}
\description{
Clean Service Route Numbers
}
