% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_interagency_gtfs.R
\name{combine_interagency_gtfs}
\alias{combine_interagency_gtfs}
\title{Combine GTFS Files From Interagencies}
\usage{
combine_interagency_gtfs(
  gtfs_filepath,
  save_csv = FALSE,
  save_RDS = FALSE,
  output_folder = NULL
)
}
\arguments{
\item{gtfs_filepath}{Character string in quotes. The location of the folders containing the GTFS files. This should be the parent directory, not the sub-directory.}

\item{save_csv}{T/F Do you want to save csv files of the combined GTFS tables?}

\item{save_RDS}{T/F Do you want to save an RDS object of the combined GTFS list object?}

\item{output_folder}{File path of the place to export file exports}
}
\value{
a list object with combined GTFS information. Optionally, exports to csv and RDS.
}
\description{
This function combines Metro's GTFS with other agency's GTFS (i.e. Community Transit, Pierce Transit, Sound Transit). This function is similar in function with @seealso \code{\link[=combine_gtfs]{combine_gtfs()}}.
}
