% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_to_tbird.R
\name{connect_to_tbird}
\alias{connect_to_tbird}
\title{Connect to T-BIRD Data Warehouse}
\usage{
connect_to_tbird()
}
\value{
A connection object
}
\description{
Connect to T-BIRD Data Warehouse
}
\examples{
con <- connect_to_tbird()
# Must be on KCM VPN for this to work


}
