% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_vi_trip_count.R
\name{count_trips_by_geography_title_vi}
\alias{count_trips_by_geography_title_vi}
\title{Title VI Transit Trips}
\usage{
count_trips_by_geography_title_vi(
  gtfs_object,
  project_name,
  geography = NULL,
  netplan_gtfs,
  gtfs_type,
  include_st = F,
  save_csv = T,
  save_RDS = T,
  output_folder
)
}
\arguments{
\item{gtfs_object}{A GTFS list-object available in your working environment}

\item{project_name}{Character string. Used as unique name for file exports}

\item{geography}{A sf polygon object to use as analysis geography. If null, defaults to 2020 Census Tracts}

\item{netplan_gtfs}{T/F Is the GTFS from NetPlan?}

\item{gtfs_type}{Character string. Used in output naming for exports. Typically "baseline" or "proposed".}

\item{include_st}{T/F Should ST routes be included in the trip change analysis?}

\item{save_csv}{T/F Do you want to save csv files of the combined GTFS tables?}

\item{save_RDS}{T/F Do you want to save an RDS object of the combined GTFS list object?}

\item{output_folder}{File path of the place to export file exports}
}
\value{
csv files of results
}
\description{
Title VI Transit Trips
}
