% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trips_table.R
\name{create_trips_table}
\alias{create_trips_table}
\title{Create summary table of trips and headways}
\usage{
create_trips_table(day_type, network, by_direction = TRUE, by_period = TRUE)
}
\arguments{
\item{day_type}{Character. Day of Week. Options are 'wkd', 'sat', 'sun'}

\item{network}{Character. GTFS. Options are 'baseline_gtfs', 'proposed_gtfs'}

\item{by_direction}{T/F. Do you want a breakdown by inbound/outbound direction?}

\item{by_period}{T/F. Do you want a breakdown by period?}

\item{routes}{Character. Filter by Routes}
}
\value{
Data frame of summary metrics of GTFS input (Trip Count, Span, Headway)
}
\description{
Create summary table of trips and headways
}
\examples{
baseline_gtfs <- tidytransit::read_gtfs(path =
fs::path_package( "extdata", "gtfs", "241_gtfs.zip", package = "ServicePlanningFunctions"))

spring_24_trip_table <- create_trips_table(day_type = 'wkd', network = 'baseline_gtfs', by_direction = FALSE, by_period = TRUE)
}
