% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_productivity_thresholds.R
\name{get_productivity_thresholds}
\alias{get_productivity_thresholds}
\title{Generate a dataframe of
productivity thresholds for both riders per platform hour and
passenger miles per platform mile. This function allows the user
to either return results for the whole system for a service change
or return a subset of results. The subsets are evaluated against all
routes before being returned. The function also supports both Metro
productivity period analysis and day part level analysis. Multiple
service changes are also supported.}
\usage{
get_productivity_thresholds(
  service_change,
  tbird_connection,
  period_type,
  sched_day_type_coded_num
)
}
\arguments{
\item{service_change}{Numeric. The three-digit identifier of the service change.
Can accept multiple values as a vector.}

\item{tbird_connection}{The connection object created by connect_to_tbird()}

\item{period_type}{Character. Level of time aggregation. Options are "day_part_cd" or "service_guidelines".}

\item{sched_day_type_coded_num}{Numeric. Day of the week. 0 - Weekday, 1 - Saturday, 2 - Sunday. Can accept multiples.}
}
\value{
Dataframe of productivity thresholds for specified days and service changes. Includes both riders per platform hour and
passenger miles per platform mile.
}
\description{
Generate a dataframe of
productivity thresholds for both riders per platform hour and
passenger miles per platform mile. This function allows the user
to either return results for the whole system for a service change
or return a subset of results. The subsets are evaluated against all
routes before being returned. The function also supports both Metro
productivity period analysis and day part level analysis. Multiple
service changes are also supported.
}
