% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_route_equity_scores.R
\name{get_route_equity_scores}
\alias{get_route_equity_scores}
\title{Get Route Equity
This function pulls the dp.route_classification table in TBIRD.}
\usage{
get_route_equity_scores(tbird_connection, route = NULL, service_change = 253)
}
\arguments{
\item{tbird_connection}{The connection object created by connect_to_tbird()}

\item{route}{Numeric. The route identifiers of interest. Values to be treated as characters to allow for non-numeric route identifiers.
Can accept multiple values as a vector.}

\item{service_change}{Numeric. The three-digit identifier of the service change.
Can accept multiple values as a vector.}
}
\value{
A dataframe of equity data by route and service change.
}
\description{
Get Route Equity
This function pulls the dp.route_classification table in TBIRD.
}
