% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_route_productivity.R
\name{get_route_productivity}
\alias{get_route_productivity}
\title{Get Route Productivity}
\usage{
get_route_productivity(
  service_change = 251,
  tbird_connection,
  period_type = "service_guidelines",
  sched_day_type_coded_num = c(0, 1, 2),
  filter_routes = FALSE,
  route = NULL
)
}
\arguments{
\item{service_change}{Numeric. The three-digit identifier of the service change.
Can accept multiple values as a vector.}

\item{tbird_connection}{The connection object created by connect_to_tbird()}

\item{period_type}{Character. Level of time aggregation. Options are "day_part_cd" or "service_guidelines".}

\item{sched_day_type_coded_num}{Numeric. Day of the week. 0 - Weekday, 1 - Saturday, 2 - Sunday. Can accept multiples.}

\item{filter_routes}{T/F. Do you want to return results for the entire system or for a selection of routes.}

\item{route}{Numeric. Required only if filter_routes == TRUE. The route identifiers of interest. Values to be treated as characters to allow for non-numeric route identifiers.
Can accept multiple values as a vector.}
}
\value{
Dataframe of productivity metrics for specified routes, periods and service changes. Includes both riders per platform hour and
passenger miles per platform mile.
}
\description{
get_route_productivity() generates a dataframe of
productivity metrics for both riders per platform hour and
passenger miles per platform mile. This function allows the user
to either return results for the whole system for a service change
or return a subset of results. The subsets are evaluated against all
routes before being returned. The function also supports both Metro
productivity period analysis and day part level analysis. Multiple
service changes are also supported.
}
