% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_route_ridership_by_area.R
\name{get_route_ridership_by_area}
\alias{get_route_ridership_by_area}
\title{Get Route Ridership by Area}
\usage{
get_route_ridership_by_area(
  area,
  gtfs_date,
  service_change_num,
  tbird_connection,
  return_type,
  sched_day_type_coded_num,
  time_period = c("AM", "PM", "MID", "XEV", "XNT"),
  activity_type = "ons",
  data_source
)
}
\arguments{
\item{area}{Character. The name of the LOCUS district of interest. Can accept multiples.}

\item{gtfs_date}{'YYYY-MM-DD' Date of GTFS file to use for stop locations. Defaults to nearest dataset released before or on specified date.}

\item{service_change_num}{Numeric. The three-digit identifier of the service change.
Can accept multiple values as a vector if you are returning a table. For maps, select one service change at a time.}

\item{tbird_connection}{The connection object created by connect_to_tbird()}

\item{return_type}{Character. Specify what you want to see. Options are "table" and "interactive_map".}

\item{sched_day_type_coded_num}{Numeric. 0 = Weekday, 1 = Saturday, 2 = Sunday.}

\item{time_period}{Character. AM, PM, MID, XEV. XNT.}

\item{activity_type}{Character. ons - Average Daily Boarding, offs - Average Daily Alightings, avg_lod - Average Max Load.}

\item{data_source}{Character. Specify which areas to show. Options are 'LOCUS' or 'King County Council Districts'}
}
\value{
Interactive map object or dataframe of route ridership summed by selected periods.
}
\description{
This function returns either an interactive map of route  ridership for the selected area or a dataframe showing ridership data for routes that serve the selected area.
}
