% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stop_ridership.R
\name{get_stop_ridership}
\alias{get_stop_ridership}
\title{Get Stop Ridership from T-BIRD
Export stop level ons, offs and load from DP.STOP_ACTIVITY_DETAIL}
\usage{
get_stop_ridership(service_change_num, route, stop_id, tbird_connection)
}
\arguments{
\item{service_change_num}{Numeric. The three-digit identifier of the service change.
Can accept multiple values as a vector.}

\item{route}{Numeric. The route identifiers of interest or "All" to return all routes. Values to be treated as characters to allow for non-numeric route identifiers.
Can accept multiple values as a vector.}

\item{stop_id}{A vector of stop IDs or "All" if all stops on a route are of interest.}

\item{tbird_connection}{The connection object created by connect_to_tbird()}
}
\value{
A dataframe of stop activity for selected routes, service changes and stops.
}
\description{
Get Stop Ridership from T-BIRD
Export stop level ons, offs and load from DP.STOP_ACTIVITY_DETAIL
}
