% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stops_by_area.R
\name{get_stops_by_area}
\alias{get_stops_by_area}
\title{Get Stops by Area}
\usage{
get_stops_by_area(area, gtfs_date, tbird_connection, return_type, data_source)
}
\arguments{
\item{area}{Character. The name of the LOCUS district of interest. Can accept multiples.}

\item{gtfs_date}{'YYYY-MM-DD' Date of GTFS file to use for stop locations. Defaults to nearest dataset released before or on specified date. Can accept multiples.}

\item{tbird_connection}{The connection object created by connect_to_tbird()}

\item{return_type}{Character. Specify what you want to see. Options are "table" and "interactive_map".}

\item{data_source}{Character. Specify which areas to show. Options are 'LOCUS' or 'King County Council Districts'}
}
\value{
Interactive map object or table of stops in selected area.
}
\description{
Returns a map or table of stops within a specified area. Areas are based on the LOCUS districts shapefile. Use show_areas() to see a map of all available areas.
}
