% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_title_vi_classification.R
\name{get_title_vi_classification}
\alias{get_title_vi_classification}
\title{Get Title VI Classification}
\usage{
get_title_vi_classification(acs_year, geography, state = "WA", county = "King")
}
\arguments{
\item{acs_year}{Numeric. Year of data to return for American Community Survey}

\item{geography}{A sf polygon object to use as analysis geography. If null, defaults to 2020 Census Tracts}

\item{state}{Character. Two character state code for ACS data. Should match geography object. Defaults to "WA"}

\item{county}{Character. Full name of county for ACS data. Should match geography object. Defaults to "King"}
}
\value{
sf object with fields indicating whether or not there is a significant bipoc or low income population in the tract.
}
\description{
Get Title VI Classification
}
