% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_calendar_full_dates.R
\name{gtfs_calendar_full_dates}
\alias{gtfs_calendar_full_dates}
\title{Determine Service Dates for GTFS}
\usage{
gtfs_calendar_full_dates(
  calendar,
  calendar_dates = NULL,
  netplan_gtfs = FALSE,
  designated_start_date = NULL,
  designated_end_date = NULL
)
}
\arguments{
\item{calendar}{Dataframe. Should be the calendar.txt file from the GTFS you are working with}

\item{calendar_dates}{Dataframe. Should be the calendar_dates.txt file from the GTFS you are working with}

\item{netplan_gtfs}{T/F Is the GTFS you are using exported from NetPlan?}

\item{designated_start_date}{Numeric. YYYYMMDD defines start_date in calendar table, as NetPlan does not export it. Required if not already defined in calendar.txt.}

\item{designated_end_date}{Numeric. YYYYMMDD defines end_date in calendar table, as NetPlan does not export it. Required if not already defined in calendar.txt.}
}
\value{
Dataframe of dates that the service_id will be in service.
}
\description{
Determine Service Dates for GTFS
}
\examples{
spring_24_gtfs <- tidytransit::read_gtfs(path =
fs::path_package( "extdata", "gtfs", "241_gtfs.zip", package = "ServicePlanningFunctions"))
full_cal <- gtfs_calendar_full_dates(calendar = spring_24_gtfs$calendar,
 calendar_dates = spring_24_gtfs$calendar_dates, netplan_gtfs = FALSE)

netplan_sept24_gtfs <- tidytransit::read_gtfs(path =
 fs::path_package( "extdata", "gtfs",
 "SEPT24_TRIP_GTFS.zip",
  package = "ServicePlanningFunctions"))

netplan_full_cal <- gtfs_calendar_full_dates( calendar = netplan_sept24_gtfs$calendar,
                                           calendar_dates = netplan_sept24_gtfs$calendar_dates,
                                          netplan_gtfs = TRUE, designated_start_date = 20240901,
                                           designated_end_date = 20250330)

}
