% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kcm_colors_ref}
\alias{kcm_colors_ref}
\title{Colors Legend}
\format{
A data frame with 24 rows and 4 variables:
\describe{
\item{Function}{Function to use colors}
\item{Variable}{Variable Type}
\item{Value}{Value}
\item{Color}{HEX Color Code}
}
}
\usage{
kcm_colors_ref
}
\description{
A dataset containing the mapping of values to colors.
}
\keyword{datasets}
