% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_route_by_service_change.R
\name{plot_route_by_service_change}
\alias{plot_route_by_service_change}
\title{Generate plot of ons, offs, and load by Route and Service Change from get_trip_ridership()}
\usage{
plot_route_by_service_change(
  dataframe,
  service_change_num,
  route,
  day = c("Weekday", "Saturday", "Sunday"),
  time_period = c("AM", "PM", "MID", "XEV", "XNT"),
  activity_type = "ons"
)
}
\arguments{
\item{dataframe}{Dataframe. Output from get_trip_ridership().}

\item{service_change_num}{Numeric. The three-digit identifier of the service change. Can accept multiple values as a vector.}

\item{route}{Numeric. The route identifiers of interest. Values to be treated as characters to allow for non-numeric route identifiers. Can accept multiple values as a vector.}

\item{day}{Character. Day of the week. Weekday, Saturday, Sunday.}

\item{time_period}{Character. AM, PM, MID, XEV. XNT.}

\item{activity_type}{Character. ons - Average Daily Boarding, offs - Average Daily Alightings, avg_lod - Average Max Load.}
}
\description{
Generate plot of ons, offs, and load by Route and Service Change from get_trip_ridership()
}
