% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_route_ridership_by_area.R
\name{plot_route_ridership_by_area}
\alias{plot_route_ridership_by_area}
\title{Plot Route Ridership within Area
This function filters routes to routes with stops in a geographic area, determines ridership on the route and generates a plot. Ridership is for the entire route, not the areas used as a filter.}
\usage{
plot_route_ridership_by_area(
  area,
  gtfs_date,
  service_change_num,
  tbird_connection,
  sched_day_type_coded_num,
  time_period = c("AM", "PM", "MID", "XEV", "XNT"),
  activity_type = "ons",
  data_source,
  x_axis
)
}
\arguments{
\item{area}{Character. The name of the LOCUS district of interest. Can accept multiples.}

\item{gtfs_date}{'YYYY-MM-DD' Date of GTFS file to use for stop locations. Defaults to nearest dataset released before or on specified date.}

\item{service_change_num}{Numeric. The three-digit identifier of the service change.
Can accept multiple values as a vector if you are returning a table. For maps, select one service change at a time.}

\item{tbird_connection}{The connection object created by connect_to_tbird()}

\item{sched_day_type_coded_num}{Numeric. 0 = Weekday, 1 = Saturday, 2 = Sunday.}

\item{time_period}{Character. AM, PM, MID, XEV. XNT.}

\item{activity_type}{Character. ons - Average Daily Boarding, offs - Average Daily Alightings, avg_lod - Average Max Load.}

\item{data_source}{Character. Specify which areas to show. Options are 'LOCUS' or 'King County Council Districts'}

\item{x_axis}{Character. Only required if returning a plot. Options are 'neighborhood', 'route', 'period', 'hour'.}
}
\value{
Interactive map object or dataframe of route ridership summed by selected periods.

A ggplot2 plot of route ridership.
}
\description{
Plot Route Ridership within Area
This function filters routes to routes with stops in a geographic area, determines ridership on the route and generates a plot. Ridership is for the entire route, not the areas used as a filter.
}
