% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stop_ridership_by_area.R
\name{plot_stop_ridership_by_area}
\alias{plot_stop_ridership_by_area}
\title{ggplot engine for stop-level ridership plots by area}
\usage{
plot_stop_ridership_by_area(
  area,
  gtfs_date,
  service_change_num,
  route = "All",
  tbird_connection,
  time_period = c("AM", "PM", "MID", "XEV", "XNT"),
  activity_type = "ons",
  data_source,
  x_axis
)
}
\arguments{
\item{area}{Character. The name of the LOCUS or KC district of interest. Can accept multiples.}

\item{gtfs_date}{'YYYY-MM-DD' Date of GTFS file to use for stop locations. Defaults to nearest dataset released before or on specified date. Can accept multiples.}

\item{service_change_num}{Numeric. The three-digit identifier of the service change. Can accept multiple values as a vector.}

\item{route}{Numeric. The route identifiers of interest. Values to be treated as characters to allow for non-numeric route identifiers. Can accept multiple values as a vector.}

\item{tbird_connection}{The connection object created by connect_to_tbird()}

\item{time_period}{Character. AM, PM, MID, XEV. XNT.}

\item{activity_type}{Character. ons - Average Daily Boarding, offs - Average Daily Alightings, rider - Total Stop Activity}

\item{data_source}{Character. Specify which areas to show. Options are 'LOCUS' or 'King County Council Districts'}

\item{x_axis}{Character. Grouping variable based on columns found in output from get_stop_ridership(). period, hour, route, route_name.}
}
\value{
A ggplot2 plot showing ons, offs or total ridership at stops based on selected parameters.
}
\description{
Generate plot of ons, offs, and load by Select Variable and Service Change for geograpic areas
}
