% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trip_crosstab.R
\name{plot_trip_crosstab}
\alias{plot_trip_crosstab}
\title{Generate plot of ons, offs, and load}
\usage{
plot_trip_crosstab(
  dataframe,
  service_change_num,
  route,
  day = c("Weekday", "Saturday", "Sunday"),
  time_period = c("AM", "PM", "MID", "XEV", "XNT"),
  x_axis = "period",
  activity_type = "ons",
  split_by_day_type = FALSE,
  color_palette = "viridis",
  color_palette_direction = 1
)
}
\arguments{
\item{dataframe}{Dataframe. Output from get_trip_ridership().}

\item{service_change_num}{Numeric. The three-digit identifier of the service change. Can accept multiple values as a vector.}

\item{route}{Numeric. The route identifiers of interest. Values to be treated as characters to allow for non-numeric route identifiers. Can accept multiple values as a vector.}

\item{day}{Character. Day of the week. Weekday, Saturday, Sunday.}

\item{time_period}{Character. AM, PM, MID, XEV. XNT.}

\item{x_axis}{Character. Grouping variable based on columns found in output from get_trip_ridership(). day, period, hour, route, route_name.}

\item{activity_type}{Character. ons - Average Daily Boarding, offs - Average Daily Alightings, avg_lod - Average Max Load,
weekly_trips - Total of  Weekly Trips on selected routes, day_type_trips - Count of Trips by day type on selected routes}

\item{split_by_day_type}{T/F Do you want to show the plots by day side-by-side? Defaults to false}

\item{color_palette}{Character.A character string indicating the color map option to use. Eight options are available: "magma", "inferno", "plasma",
"viridis", "cividis", "rocket" , "mako"  or "turbo" .
@param color_palette_direction Numeric.     Sets the order of colors in the scale. If 1, the default,
colors are as output by viridis_pal. If -1, the order of colors is reversed.}
}
\value{
ggplot2 plot of ons, offs, and load by Select Variable and Service Change from get_trip_ridership()
}
\description{
Create plots of trip and route ridership by Select Variable and Service Change from get_trip_ridership()
}
