% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_water.R
\name{remove_water}
\alias{remove_water}
\title{Remove water bodies from polygon}
\usage{
remove_water(polygon, state_code = "WA", county_code = "King", crs = 2926)
}
\arguments{
\item{polygon}{sf object representing your polygon shapefile.}

\item{state_code}{The two-letter state code for the location in quotes. Washington is "WA".}

\item{county_code}{The name of the county your polygons are in. Used to fetch the water bodies in the area.}

\item{crs}{The four digit CRS (AKA EPSG) code for the projection system your data is in. King County's is 2926. Web Mercator is 4326}
}
\value{
sf object with erased water geometries through spatial intersection.
}
\description{
Remove water bodies from polygon
}
