% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_objects.R
\name{save_objects}
\alias{save_objects}
\title{Save Objects from R Environment to Disc}
\usage{
save_objects(object_name, file_location)
}
\arguments{
\item{object_name}{Character. The name of the object you want to save.}

\item{file_location}{The folder location on your computer that you want to write to. Currently requires full file path.}
}
\value{
Returns an RDS object containing the R object from the global environment.
}
\description{
Save Objects from R Environment to Disc
}
\examples{

a <- as.data.frame("letters") #create sample data
b<- as.data.frame(c(1,2,3,4))
object_list <- ls() # create character vector of all objects in environment
#remove any you don't want written to disk
object_list <- object_list[! object_list \%in\% c("object_list")]
#use purrr to iterate
purrr:::map(object_list,
file_location ="C:/Users/mgaughan/OneDrive - King County/Documents/test", save_objects)
}
