% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ServicePlanningFunctions-package.R
\docType{package}
\name{ServicePlanningFunctions-package}
\alias{ServicePlanningFunctions}
\alias{ServicePlanningFunctions-package}
\title{ServicePlanningFunctions: Functions to Make the Analysis of KCM Service Planning Analysis Easier}
\description{
This is a collection of functions that come up repeatedly in transit service planning analysis. There are also some functions that just make repetitive tasks easier. Check function documentation for specifics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/melissa-gaughan/ServicePlanningFunctions}
  \item Report bugs at \url{https://github.com/melissa-gaughan/ServicePlanningFunctions/issues}
}

}
\author{
\strong{Maintainer}: Melissa Gaughan \email{mgaughan@kingcounty.gov}

Authors:
\itemize{
  \item Melissa Gaughan \email{mgaughan@kingcounty.gov}
  \item Michael Shin \email{mshin@kingcounty.gov}
}

}
\keyword{internal}
