% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_kcm.R
\name{style_kcm}
\alias{style_kcm}
\title{Title}
\usage{
style_kcm(textsize = NULL)
}
\arguments{
\item{textsize}{Specifies the font size to use. Defaults to default size. "markdown" to increase font size for pdf/png/jpeg exports.}
}
\description{
Title
}
\examples{
ggplot(mtcars, aes(x = wt, y = factor(cyl), color = factor(cyl))) +
geom_boxplot() +
facet_grid(. ~ gear) +
style_kcm()
}
