% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_kcm_colors.R
\name{style_kcm_colors}
\alias{style_kcm_colors}
\title{Use Designated KCM Color Scheme for Plots}
\usage{
style_kcm_colors()
}
\value{
utility list of categories and default colors
}
\description{
Use Designated KCM Color Scheme for Plots
}
\examples{
#ggplot(mtcars, aes(x = wt, y = factor(cyl), color = factor(cyl))) +
#geom_boxplot() +
#facet_grid(. ~ gear) +
#style_kcm() +
#style_kcm_colors()
}
