% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_kcm_colors_default.R
\name{style_kcm_colors_default}
\alias{style_kcm_colors_default}
\title{Use Default KCM Color Scheme for Plots}
\usage{
style_kcm_colors_default()
}
\value{
utility list of default Metro colors
}
\description{
Use Default KCM Color Scheme for Plots
}
\examples{
#ggplot(mtcars, aes(x = wt, y = factor(cyl), color = factor(cyl))) +
#geom_boxplot() +
#facet_grid(. ~ gear) +
#style_kcm() +
#style_kcm_colors()
}
