% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_kcm_flip.R
\name{style_kcm_flip}
\alias{style_kcm_flip}
\title{Title}
\usage{
style_kcm_flip(textsize = NULL)
}
\arguments{
\item{textsize}{Specifies the font size to use. Defaults to default size. "markdown" to increase font size for pdf/png/jpeg exports.}
}
\description{
Title
}
\examples{
ggplot(mtcars, aes(x = factor(cyl), y = wt, color = factor(cyl))) +
geom_boxplot() +
facet_grid(. ~ gear) +
style_kcm_flip()
}
