% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_ServicePlanningFunctions.R
\name{update_ServicePlanningFunctions}
\alias{update_ServicePlanningFunctions}
\title{Install/Update ServicePlanningFunctions Package}
\usage{
update_ServicePlanningFunctions()
}
\value{
updates ServicePlanningFunctions package. Returns error if not successful.
}
\description{
Install/Update ServicePlanningFunctions Package
}
