% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_gtfs.R
\name{combine_gtfs}
\alias{combine_gtfs}
\title{Combine GTFS Files From NetPlan}
\usage{
combine_gtfs(
  gtfs_filepath,
  designated_start_date,
  designated_end_date,
  save_csv = FALSE,
  save_RDS = FALSE,
  output_folder = NULL
)
}
\arguments{
\item{gtfs_filepath}{Character string in quotes. The location of the folders containing the GTFS files. This should be the parent directory, not the sub-directory.}

\item{designated_start_date}{Numeric. YYYYMMDD defines start_date in calendar table, as NetPlan does not export it.}

\item{designated_end_date}{Numeric. YYYYMMDD defines end_date in calendar table, as NetPlan does not export it.}

\item{save_csv}{T/F Do you want to save csv files of the combined GTFS tables?}

\item{save_RDS}{T/F Do you want to save an RDS object of the combined GTFS list object?}

\item{output_folder}{File path of the place to export file exports}
}
\value{
a list object with combined GTFS information. Optionally, exports to csv and RDS.
}
\description{
Combine GTFS Files From NetPlan
}
\examples{
#create filepath for the folder that holds the GTFS folder sets to be combined.
 #Each GTFS needs its own subfolder
gtfs_folder <- fs::path_package( "extdata", "combine_gtfs", package = "ServicePlanningFunctions")
# set start and end dates for service period, pass filepath to function.
madison_st_area_gtfs <- combine_gtfs(gtfs_filepath = gtfs_folder, designated_start_date = 20240901,
 designated_end_date = 20250330, save_csv = FALSE, save_RDS = FALSE, output_folder = NULL)
}
