% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trip_ridership.R
\name{get_trip_ridership}
\alias{get_trip_ridership}
\title{Get Trip Ridership from T-BIRD Export trip level ons, offs, and load from DP.VW_TRIP_SUMMARY}
\usage{
get_trip_ridership(
  service_change_num,
  route,
  sched_day_type_coded_num,
  tbird_connection
)
}
\arguments{
\item{service_change_num}{Numeric. The three-digit identifier of the service change. Can accept multiple values as a vector.}

\item{route}{Numeric. The route identifiers of interest. Values to be treated as characters to allow for non-numeric route identifiers. Can accept multiple values as a vector.}

\item{sched_day_type_coded_num}{Numeric. Day of the week. 0 - Weekday, 1 - Saturday, 2 - Sunday.}

\item{tbird_connection}{The connection object created by connect_to_tbird()}
}
\value{
Dataframe of trip level ons, offs, and load from DP.VW_TRIP_SUMMARY
}
\description{
Get Trip Ridership from T-BIRD Export trip level ons, offs, and load from DP.VW_TRIP_SUMMARY
}
