% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_productivity_distribution.R
\name{plot_productivity_distribution}
\alias{plot_productivity_distribution}
\title{ggplot engine for route-level productivity distribution plots}
\usage{
plot_productivity_distribution(
  service_change,
  tbird_connection,
  svc_family = NULL,
  sched_day_type_coded_num = 0,
  period_type = "day_part_cd",
  period = "DAY",
  route = NULL,
  route_gain = NULL,
  route_maintain = NULL,
  route_lose = NULL,
  activity_type = "rides_per_platform_hour",
  binwidth = NULL,
  point_size = 20,
  label_size = 4,
  style_size = NULL
)
}
\arguments{
\item{service_change}{Numeric. The three-digit identifier of the service change.}

\item{tbird_connection}{The connection object created by connect_to_tbird()}

\item{svc_family}{Character. Service Family. Options are 'Urban', 'Suburban', and 'ST'.}

\item{sched_day_type_coded_num}{Numeric. Day of the week. 0 - Weekday, 1 - Saturday, 2 - Sunday.}

\item{period_type}{Character. Level of time aggregation. Options are "day_part_cd", or "service_guidelines". Defaults to 'day_part_cd'.}

\item{period}{Character. Time Period. For "day_part_cd" period_type, options are AM, PM, MID, XEV, XNT, DAY. DAY represents All Day. For "service_guidelines" period_type, options are Peak, Off-Peak, Night. Defaults to 'DAY'.}

\item{route}{Numeric. The route identifiers of interest to highlight and label. Values to be treated as characters to allow for non-numeric route identifiers. Can accept multiple values as a vector. Defaults to show all routes. If any of the route_gain, route_maintain, or route_lose paramenters are populated, this parameter is ignored.}

\item{route_gain}{Numeric. Routes to color as gaining service. Overrides route parameter if used.}

\item{route_maintain}{Numeric. Routes to color as maintaining service. Overrides route parameter if used.}

\item{route_lose}{Numeric. Routes to color as losing service. Overrides route parameter if used.}

\item{activity_type}{Character. rides_per_platform_hour - Rides per Platform Hour, psngr_miles_per_platform_mile - Passenger Miles per Platform Mile}

\item{binwidth}{Numeric. Specifies bin width. Decrease if dots are too close to each other vertically. Increase if dots are too close to each other horizontally. Defaults to 2 for rides per platform hour and 1 for passenger miles per platform miles.}

\item{point_size}{Numeric. The diameter of the dots. Adjust so that dots do not overlap one another. Interacts with binwidth.}

\item{label_size}{Numeric. The font size of the dot labels. Adjust so that labels fit within each dot.}

\item{style_size}{Numeric. Specifies the sizing style to use with style_kcm(). Defaults to default size. "large" to increase sizing style for pdf/png/jpeg/markdown exports.}
}
\value{
A ggplot2 plot showing distribution dotplot of route productivity metrics.
}
\description{
Generate a distribution dotplot of route productivity metrics from get_route_productivity(). Rides per Platform Hour or Passenger Miles per Platform Mile.
}
\examples{
con <- connect_to_tbird()

# Generate Rides per Platform Hour distribution dotplot for Fall 2025 Service Change
plot_productivity_distribution(service_change = 253, tbird_connection = con)

# Highlight Routes based on Service Gain/Loss
plot_productivity_distribution(service_change = 253, tbird_connection = con, route_gain = c(8), route_maintain = c(1, 2), route_lose = c(107, 673))

# Generate Passenger Miles per Platform Mile distribution dotplot for Fall 2025 Service Change highlighting RapidRide routes only.
# Sizing parameters formatted for png download size.
plot_productivity_distribution(service_change = 253,
 tbird_connection = con,
 route = c(671, 672, 673, 674, 675, 676, 677, 678),
 activity_type = 'psngr_miles_per_platform_mile',
 binwidth = 1,
 point_size = 6.5,
 label_size = 8,
 style_size = 'large')

ggplot2::ggsave("RapidRide Passenger Miles per Platform Mile.png", width = 12.5, height = 6.9, units = "in")
}
