% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stop_crosstab.R
\name{plot_stop_crosstab}
\alias{plot_stop_crosstab}
\title{ggplot engine for stop level ridership plots
Generate plot of ons, offs, and load by Select Variable and Service Change from get_stop_ridership()}
\usage{
plot_stop_crosstab(
  dataframe,
  service_change_num,
  route,
  time_period = c("AM", "PM", "MID", "XEV", "XNT"),
  x_axis,
  activity_type = "ons"
)
}
\arguments{
\item{dataframe}{Dataframe. Output from get_trip_ridership().}

\item{service_change_num}{Numeric. The three-digit identifier of the service change. Can accept multiple values as a vector.}

\item{route}{Numeric. The route identifiers of interest. Values to be treated as characters to allow for non-numeric route identifiers. Can accept multiple values as a vector.}

\item{time_period}{Character. AM, PM, MID, XEV. XNT.}

\item{x_axis}{Character. Grouping variable based on columns found in output from get_stop_ridership(). period, hour, route, route_name.}

\item{activity_type}{Character. ons - Average Daily Boarding, offs - Average Daily Alightings, rider - Total Stop Activity}
}
\value{
A ggplot2 plot of ons, offs, and load by Select Variable and Service Change from get_stop_ridership()
}
\description{
ggplot engine for stop level ridership plots
Generate plot of ons, offs, and load by Select Variable and Service Change from get_stop_ridership()
}
