% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_areas.R
\name{show_areas}
\alias{show_areas}
\title{Show Areas
Returns either a table of available district names or a map showing the district boundaries.
Based on LOCUS districts. Designed to be used with get_stops_by_area() and get_routes_by_area().}
\usage{
show_areas(return_type = "interactive_map")
}
\arguments{
\item{return_type}{Character. Specify what you want to see. Options are "table" and "interactive_map".}
}
\value{
A list of area names or an interactive map of area boundaries.
}
\description{
Show Areas
Returns either a table of available district names or a map showing the district boundaries.
Based on LOCUS districts. Designed to be used with get_stops_by_area() and get_routes_by_area().
}
