% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_erase.R
\name{st_erase}
\alias{st_erase}
\title{st_erase}
\usage{
st_erase(x, y)
}
\arguments{
\item{x}{shapefile to remove shapes from}

\item{y}{Shapefile of shapes to be removed}
}
\value{
returns union of the shapes, with y geometries removed
}
\description{
st_erase
}
